fix: ease coercion for account feature overrides
BREAKING CHANGES: The `feature_overrides` field is currently coerced
into an array of strings. This data structure is optimized for policy
resolution in the service but does not represent the actual data that is
given when maintaining the account or what is stored in the db.

Easing this coercion is preparation for changing the actual data
returned which will happen in a future PR.